Update model field

+++ workouts/models.py
@@ -44,7 +44,7 @@ class WorkoutLog(models.Model):
 
     log_name = models.ForeignKey(WorkoutTemplate, on_delete=models.CASCADE)
     exercise_type = models.ForeignKey(Exercises, on_delete=models.CASCADE)
-    weight = models.PositiveIntegerField(default=10)
+    weight = models.DecimalField(default=10.0, max_digits=3, decimal_places=1)
     sets = models.PositiveIntegerField(default=3)
     reps = models.PositiveIntegerField(default=5)
     note = models.TextField()
